regroup all sdk types in the client dir
