+++ backend/chainlit/client/cloud.py
@@ -5,21 +5,24 @@ from typing import Dict, Optional, Union
 import aiohttp
 from chainlit.config import config
 from chainlit.logger import logger
-from chainlit.types import AppUser, ConversationFilter, Pagination, PersistedAppUser
 
 from .base import (
+    AppUser,
     ChainlitGraphQLClient,
     ConversationDict,
+    ConversationFilter,
     ElementDict,
     MessageDict,
     PageInfo,
     PaginatedResponse,
+    Pagination,
+    PersistedAppUser,
 )
 
 
 class ChainlitCloudClient(ChainlitGraphQLClient):
-    def __init__(self, api_key: str):
-        super().__init__(api_key=api_key)
+    def __init__(self, chainlit_server: str, api_key: str):
+        super().__init__(chainlit_server=chainlit_server, api_key=api_key)
 
     async def create_app_user(self, app_user: AppUser) -> Optional[PersistedAppUser]:
         mutation = """
@@ -451,5 +454,6 @@ chainlit_client = None  # type: Optional[ChainlitCloudClient]
 
 if config.data_persistence:
     chainlit_client = ChainlitCloudClient(
-        api_key=os.environ.get("CHAINLIT_API_KEY", "")
+        chainlit_server=config.chainlit_server,
+        api_key=os.environ.get("CHAINLIT_API_KEY", ""),
     )
