+++ backend/chainlit/element.py
@@ -6,11 +6,10 @@ from typing import Any, ClassVar, Dict, List, Optional, TypeVar, Union, cast
 
 import aiofiles
 import filetype
-from chainlit.client.base import ElementDict
+from chainlit.client.base import ElementDict, ElementDisplay, ElementSize, ElementType
 from chainlit.client.cloud import ChainlitCloudClient, chainlit_client
 from chainlit.context import context
 from chainlit.telemetry import trace_event
-from chainlit.types import ElementDisplay, ElementSize, ElementType
 from pydantic.dataclasses import Field, dataclass
 
 mime_types = {
