+++ backend/chainlit/types.py
@@ -1,19 +1,15 @@
 from enum import Enum
-from typing import Any, Dict, List, Literal, Optional, TypedDict, Union
+from typing import Dict, List, Literal, TypedDict, Union
 
+from chainlit.client.base import ConversationFilter, Pagination
 from chainlit.prompt import Prompt
 from dataclasses_json import DataClassJsonMixin
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 from pydantic.dataclasses import dataclass
 
 InputWidgetType = Literal[
     "switch", "slider", "select", "textinput", "tags", "numberinput"
 ]
-ElementType = Literal[
-    "image", "avatar", "text", "pdf", "tasklist", "audio", "video", "file"
-]
-ElementDisplay = Literal["inline", "side", "page"]
-ElementSize = Literal["small", "medium", "large"]
 
 
 @dataclass
@@ -64,17 +60,6 @@ class DeleteConversationRequest(BaseModel):
     conversationId: str
 
 
-class Pagination(BaseModel):
-    first: int
-    cursor: Any
-
-
-class ConversationFilter(BaseModel):
-    feedback: Optional[Literal[-1, 0, 1]]
-    username: Optional[str]
-    search: Optional[str]
-
-
 class GetConversationsRequest(BaseModel):
     pagination: Pagination
     filter: ConversationFilter
@@ -83,28 +68,3 @@ class GetConversationsRequest(BaseModel):
 class Theme(str, Enum):
     light = "light"
     dark = "dark"
-
-
-Role = Literal["USER", "ADMIN", "OWNER", "ANONYMOUS"]
-Provider = Literal["credentials", "header", "github", "google", "azure-ad"]
-
-
-# Used when logging-in a user
-@dataclass
-class AppUser(DataClassJsonMixin):
-    username: str
-    role: Role = "USER"
-    tags: List[str] = Field(default_factory=list)
-    image: Optional[str] = None
-    provider: Optional[Provider] = None
-
-
-@dataclass
-class PersistedAppUserFields:
-    id: str
-    createdAt: int
-
-
-@dataclass
-class PersistedAppUser(AppUser, PersistedAppUserFields):
-    pass
